+++ NEWS.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+## Added
+- `plot_difference()` to plot difference smooths
+
 ## [1.4.0] - 2018-07-03
 ### Added
 - `split` argument for separating columns (useful for interactions)